# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "K7 E-Library API",
        "description": "Digital library management system API for educational institutions",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8747/api/v1"
        }
    ],
    "paths": {
...
        "/admin/reports/downloads": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Download statistics",
                "operationId": "9d827a37868ae5af5eb759bc4d72567e",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "daily",
                                "weekly",
                                "monthly"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Download statistics"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/reports/popular": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Most popular content",
                "operationId": "16397fb6f231ac3795e37cb26e04be6d",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Popular content"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/reports/activity": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "User activity logs",
                "operationId": "b186650bb748e896fd43441ac7cb7ce2",
                "responses": {
                    "200": {
                        "description": "Activity logs"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List all users",
                "operationId": "33c1239816ff96dc74d67a3a1bb46aa3",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create user",
                "operationId": "24efa80f1f4f75906e8ee861c676dae8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "super_admin",
                                            "admin",
                                            "faculty",
                                            "student",
                                            "guest"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/users/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update user",
                "operationId": "4b5c6dc35146585c62c0663477c29d9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete user",
                "operationId": "f651b06aa7606d31d1a54bbde55318e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}